feat: first praise -> activate
